trading_bot/requirements.txt
python-dotenv
ccxt
pandas
numpy
ta
matplotlib
python-telegram-bot
schedule
python-dateutil
retry
pytest
pytest-mock
boto3